+++ tableau_utilities/core/setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup
 
-with open('README.md') as f:
+with open('../../README.md') as f:
     readme = f.read()
 
 setup(
