+++ tableau_utilities/core/tableau_server/refresh.py
@@ -1,6 +1,5 @@
 import tableau_utilities.tableau_server.tableau_server_objects as tso
-from requests import Session
-from tableau_utilities.tableau_server.static import transform_tableau_object
+from tableau_utilities.core.tableau_server.static import transform_tableau_object
 from tableau_utilities.tableau_server.base import Base
