+++ tableau_utilities/core/tableau_server/static.py
